services/equipo-connect/requirements.txt
watchdog==5.0.3
httpx==0.27.2
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
